+++ packages/react/framework7-react.js
@@ -1,5 +1,5 @@
 /**
- * Framework7 React 3.0.0-beta.15
+ * Framework7 React 3.0.0-beta.16
  * Build full featured iOS & Android apps using Framework7 & React
  * http://framework7.io/react/
  *
@@ -7,7 +7,7 @@
  *
  * Released under the MIT License
  *
- * Released on: June 27, 2018
+ * Released on: July 1, 2018
  */
 
 (function (global, factory) {
@@ -4378,6 +4378,7 @@
       var value = props.value;
       var name = props.name;
       var checked = props.checked;
+      var defaultChecked = props.defaultChecked;
       var readonly = props.readonly;
       var disabled = props.disabled;
       var required = props.required;
@@ -4480,6 +4481,7 @@
             value: value,
             name: name,
             checked: checked,
+            defaultChecked: defaultChecked,
             readOnly: readonly,
             disabled: disabled,
             required: required,
@@ -4716,6 +4718,7 @@
     inlineLabel: Boolean,
     checkbox: Boolean,
     checked: Boolean,
+    defaultChecked: Boolean,
     radio: Boolean,
     name: String,
     value: [String, Number, Array],
@@ -4903,6 +4906,7 @@
       var checkbox = props.checkbox;
       var radio = props.radio;
       var checked = props.checked;
+      var defaultChecked = props.defaultChecked;
       var name = props.name;
       var value = props.value;
       var readonly = props.readonly;
@@ -4932,6 +4936,7 @@
           accordionItem: accordionItem,
           checkbox: checkbox,
           checked: checked,
+          defaultChecked: defaultChecked,
           radio: radio,
           name: name,
           value: value,
@@ -4980,7 +4985,9 @@
           style: style,
           className: liClasses
         }, React.createElement('span', null, this.slots['default'], !this.slots.default && title));
-      } else if (isSimple) {
+      }
+
+      if (isSimple) {
         return React.createElement('li', {
           ref: function (__reactNode) {
             this$1.__reactRefs['el'] = __reactNode;
@@ -5182,6 +5189,7 @@
     checkbox: Boolean,
     radio: Boolean,
     checked: Boolean,
+    defaultChecked: Boolean,
     name: String,
     value: [String, Number, Array],
     readonly: Boolean,
@@ -6862,7 +6870,6 @@
       var style = props.style;
       var sliding = props.sliding;
       var classes = Utils.classNames(className, 'right', {
-        right: true,
         sliding: sliding
       }, Mixins.colorClasses(props));
       return React.createElement('div', {
@@ -10619,7 +10626,7 @@
   };
 
   /**
-   * Framework7 React 3.0.0-beta.15
+   * Framework7 React 3.0.0-beta.16
    * Build full featured iOS & Android apps using Framework7 & React
    * http://framework7.io/react/
    *
@@ -10627,7 +10634,7 @@
    *
    * Released under the MIT License
    *
-   * Released on: June 27, 2018
+   * Released on: July 1, 2018
    */
 
   var Plugin = {