+++ packages/vue/components/list-item.js
@@ -36,6 +36,7 @@ export default {
     checkbox: Boolean,
     radio: Boolean,
     checked: Boolean,
+    defaultChecked: Boolean,
     name: String,
     value: [String, Number, Array],
     readonly: Boolean,
@@ -97,6 +98,7 @@ export default {
       checkbox,
       radio,
       checked,
+      defaultChecked,
       name,
       value,
       readonly,
@@ -132,6 +134,7 @@ export default {
           accordionItem: accordionItem,
           checkbox: checkbox,
           checked: checked,
+          defaultChecked: defaultChecked,
           radio: radio,
           name: name,
           value: value,
@@ -182,7 +185,9 @@ export default {
           id: id
         }
       }, [_h('span', [this.$slots['default'] || [title]])]);
-    } else if (isSimple) {
+    }
+
+    if (isSimple) {
       return _h('li', {
         ref: 'el',
         style: style,